Started work on the next version of the render loop.
This version will sort based on Priority (from HIGHEST to ROCK_BOTTOM,
respectively), then on the PlugMode (preLogic to idleRender,
respectively). Each PlugMode will have a HashSet of Methods that will be
executed.

The benefits of this layout is that:
1.) I can control the priority of the method by using an Annotation
2.) I can reduce the amount of Methods being exectued by using an
Annotation
3.) Each Plugable will be executed "correctly" (currently, they are
executing per-plugable and not per method, so each plugable is getting
executed fully before the next one starts.

The issues are:
1.) I need to completely rewrite the render loop to use the new layout
2.) I need to completely rewrite the register functions to use the new
layout